--- a/README.md
+++ b/README.md
@@ -1,3 +1,16 @@
 # tmu-courses-planner
 
 Coming soon....
+
+
+# How to view
+
+To install required packages
+```bash
+pnpm install
+```
+
+To run local development servers
+```bash 
+pnpm run dev
+```